requirements.txt
Django==3.0.4
pip==20.0.2
djangorestframework==3.11.0
django-cors-headers==3.2.1
django-heroku==0.3.1
gunicorn==20.0.4 
